// types/user.ts
export interface userTypes {
    user_id: number;
    mail: string;
    password: string;
    zip_code: string;
    city: string;
    address: string;
    phone_number: string;
    is_active: boolean;
    is_pending: boolean;
    is_to_be_completed: boolean;
    role: string;
    description: string;
    avatar: string;
}